requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.5.0
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.0.0
pylint>=3.0.0
pydantic>=2.0.0
pyyaml>=6.0.0